src/requirements.txt
librosa==0.9.1
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.1.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
